requirements.txt
fastapi~=0.115.12
uvicorn~=0.34.2
motor~=3.7.1
python-dotenv~=1.0.0
pytz~=2024.1
pydantic~=2.11.4
pydantic[email]
configparser~=7.2.0